Update docker container to run as non-root user
Nginx by default run as root user. This is because it is designed to run on port 80. Updated the container image build so that it runs as nginx user and serve over port 3000
Added CI docker build pipeline where docker image is build when a pull request is created/updated. Image published when push on main or PR is merged to main